Update admin,forms,views, remove ajax from forms and views, update base with msg

+++ bookings/views.py
@@ -1,17 +1,14 @@
-from django.http import JsonResponse
 from django.shortcuts import render, redirect, get_object_or_404
 from .models import MenuItem, Table, Booking
 from .forms import BookingForm
 from django.contrib.auth.decorators import login_required
 from django.contrib import messages
-from django.utils.html import escape
 from .utils import send_confirmation_mail
+from django.core.exceptions import PermissionDenied
 
 @login_required
 def make_booking(request):
     if request.method == 'POST':
-        is_ajax = request.headers.get('x-requested-with') == 'XMLHttpRequest'
-        
         form = BookingForm(request.POST)
         if form.is_valid():
             booking = form.save(commit=False)
@@ -32,23 +29,15 @@ def make_booking(request):
             ]
 
             if not available_tables:
-                response_data = {'success': False, 'errors': {'__all__': ['No tables available for this date and time.']}}
-                if is_ajax:
-                    return JsonResponse(response_data)
-                else:
-                    messages.error(request, 'No tables available for this date and time.')
-                    return render(request, 'bookings/make_booking.html', {'form': form})
+                messages.error(request, 'No tables available for this date and time.')
+                return render(request, 'bookings/make_booking.html', {'form': form})
 
             # Find the table with the lowest capacity among the available ones
             try:
                 lowest_capacity_table = min(available_tables, key=lambda table: table.size)
             except ValueError:
-                response_data = {'success': False, 'errors': {'__all__': ['An error occurred while selecting a table. Please try again.']}}
-                if is_ajax:
-                    return JsonResponse(response_data)
-                else:
-                    messages.error(request, 'An error occurred while selecting a table. Please try again.')
-                    return render(request, 'bookings/make_booking.html', {'form': form})
+                messages.error(request, 'An error occurred while selecting a table. Please try again.')
+                return render(request, 'bookings/make_booking.html', {'form': form})
 
             booking.table = lowest_capacity_table
 
@@ -68,28 +57,17 @@ def make_booking(request):
                     context=context,
                     template_name='emails/booking_confirmation.html'
                 )
-                
-                response_data = {'success': True, 'message': 'Booking confirmed!'}
-                if is_ajax:
-                    return JsonResponse(response_data)
-                else:
-                    messages.success(request, 'Booking confirmed!')
-                    return redirect('booking_detail', booking_id=booking.id)
+
+                messages.success(request, 'Booking confirmed!')
+                return redirect('booking_detail', booking_id=booking.id)
+
             except Exception as e:
-                response_data = {'success': False, 'errors': {'__all__': [f'An error occurred while saving your booking: {escape(str(e))}']}}
-                if is_ajax:
-                    return JsonResponse(response_data)
-                else:
-                    messages.error(request, f'An error occurred while saving your booking: {e}')
+                messages.error(request, f'An error occurred while saving your booking: {e}')
 
         # If the form is not valid, re-render the template with the errors
-        response_data = {'success': False, 'errors': form.errors.as_json()}
-        if is_ajax:
-            return JsonResponse(response_data)
-        else:
-            return render(request, 'bookings/make_booking.html', {'form': form})
+        return render(request, 'bookings/make_booking.html', {'form': form})
 
-    else:  # Handle GET requests (render the form)
+    else:
         form = BookingForm()
 
     return render(request, 'bookings/make_booking.html', {'form': form})
@@ -165,6 +143,10 @@ def delete_booking(request, booking_id):
 @login_required
 def booking_detail(request, booking_id):
     booking = get_object_or_404(Booking, id=booking_id)
+    if booking.user != request.user:
+        raise PermissionDenied()
+    else:
+        booking = get_object_or_404(Booking, id=booking_id)
     return render(request, 'bookings/booking_detail.html', {'booking': booking})
 
 
@@ -175,53 +157,59 @@ def menu(request):
 @login_required
 def edit_booking(request, booking_id):
     booking = get_object_or_404(Booking, id=booking_id)
-    
-    if request.method == 'POST':
-        form = BookingForm(request.POST, instance=booking)
-        if form.is_valid():
-            form.save()
-            
+    if booking.user == request.user:
+
+
+        if request.method == 'POST':
+            form = BookingForm(request.POST, instance=booking)
+            if form.is_valid():
+                form.save()
+                
+                # Send confirmation email
+                context = {
+                    'booking': booking,
+                    'user': request.user
+                }
+                send_confirmation_mail(
+                    subject='Booking Updated',
+                    recipient_list=[request.user.email],
+                    context=context,
+                    template_name='emails/booking_update.html'
+                )
+
+                messages.success(request, 'Booking updated!')
+                return redirect('booking_detail', booking_id=booking.id)
+        else:
+            form = BookingForm(instance=booking)
+    else:
+        raise PermissionDenied()
+    return render(request, 'bookings/edit_booking.html', {'form': form, 'booking': booking})
+
+@login_required
+def delete_booking(request, booking_id):
+    booking = get_object_or_404(Booking, id=booking_id)
+    if booking.user == request.user:
+        
+        if request.method == 'POST':
+            booking.delete()
+
             # Send confirmation email
             context = {
                 'booking': booking,
                 'user': request.user
             }
             send_confirmation_mail(
-                subject='Booking Updated',
+                subject='Booking Cancelled',
                 recipient_list=[request.user.email],
                 context=context,
-                template_name='emails/booking_update.html'
+                template_name='emails/booking_cancellation.html'
             )
 
-            messages.success(request, 'Booking updated!')
-            return redirect('booking_detail', booking_id=booking.id)
+            messages.success(request, 'Booking deleted successfully.')
+            return redirect('my_bookings')
+        
     else:
-        form = BookingForm(instance=booking)
-    
-    return render(request, 'bookings/edit_booking.html', {'form': form, 'booking': booking})
-
-@login_required
-def delete_booking(request, booking_id):
-    booking = get_object_or_404(Booking, id=booking_id)
-    
-    if request.method == 'POST':
-        booking.delete()
-
-        # Send confirmation email
-        context = {
-            'booking': booking,
-            'user': request.user
-        }
-        send_confirmation_mail(
-            subject='Booking Cancelled',
-            recipient_list=[request.user.email],
-            context=context,
-            template_name='emails/booking_cancellation.html'
-        )
-
-        messages.success(request, 'Booking deleted successfully.')
-        return redirect('my_bookings')  # Redirect to a page that lists bookings or another appropriate page
-
+        raise PermissionDenied()
     return render(request, 'bookings/delete_booking.html', {'booking': booking})
 
 @login_required
